+++ overlays/firefox.nix
@@ -1,4 +1,0 @@
-self: super: {
-  firefox-bin-unwrapped = super.firefox-bin-unwrapped.overrideAttrs
-    (old: { dontFixup = assert !old ? dontFixup; true; });
-}